+++ sdk/core/azure_core/src/http/policies/mod.rs
@@ -5,9 +5,9 @@
 
 mod bearer_token_policy;
 mod client_request_id;
-mod telemetry;
+mod user_agent;
 
 pub use bearer_token_policy::BearerTokenCredentialPolicy;
 pub use client_request_id::*;
-pub use telemetry::*;
 pub use typespec_client_core::http::policies::*;
+pub use user_agent::*;