+++ sdk/keyvault/azure_security_keyvault_certificates/tests/certificate_client.rs
@@ -3,15 +3,14 @@
 
 #![cfg_attr(target_arch = "wasm32", allow(unused_imports))]
 
-use azure_core::{http::StatusCode, sleep::sleep, time::Duration, Result};
-use azure_core_test::{recorded, Recording, TestContext, TestMode, SANITIZE_BODY_NAME};
+use azure_core::{http::StatusCode, Result};
+use azure_core_test::{recorded, ErrorKind, TestContext, TestMode, SANITIZE_BODY_NAME};
 use azure_security_keyvault_certificates::{
     models::{
-        CertificateOperation, CertificatePolicy, CreateCertificateParameters, CurveName,
-        IssuerParameters, KeyProperties, KeyType, UpdateCertificatePropertiesParameters,
-        X509CertificateProperties,
+        CertificatePolicy, CreateCertificateParameters, CurveName, IssuerParameters, KeyProperties,
+        KeyType, UpdateCertificatePropertiesParameters, X509CertificateProperties,
     },
-    CertificateClient, CertificateClientOptions, ResourceExt as _, ResourceId,
+    CertificateClient, CertificateClientExt as _, CertificateClientOptions, ResourceExt as _,
 };
 use azure_security_keyvault_keys::{
     models::{SignParameters, SignatureAlgorithm},
@@ -22,17 +21,18 @@ use futures::TryStreamExt;
 use openssl::sha::sha256;
 use std::{collections::HashMap, sync::LazyLock};
 
-static DEFAULT_POLICY: LazyLock<CertificatePolicy> = LazyLock::new(|| CertificatePolicy {
-    x509_certificate_properties: Some(X509CertificateProperties {
-        subject: Some("CN=DefaultPolicy".into()),
+static DEFAULT_CERTIFICATE_POLICY: LazyLock<CertificatePolicy> =
+    LazyLock::new(|| CertificatePolicy {
+        x509_certificate_properties: Some(X509CertificateProperties {
+            subject: Some("CN=DefaultPolicy".into()),
+            ..Default::default()
+        }),
+        issuer_parameters: Some(IssuerParameters {
+            name: Some("Self".into()),
+            ..Default::default()
+        }),
         ..Default::default()
-    }),
-    issuer_parameters: Some(IssuerParameters {
-        name: Some("Self".into()),
-        ..Default::default()
-    }),
-    ..Default::default()
-});
+    });
 
 #[recorded::test]
 async fn certificate_roundtrip(ctx: TestContext) -> Result<()> {
@@ -50,15 +50,13 @@ async fn certificate_roundtrip(ctx: TestContext) -> Result<()> {
 
     // Create a self-signed certificate.
     let body = CreateCertificateParameters {
-        certificate_policy: Some(DEFAULT_POLICY.clone()),
+        certificate_policy: Some(DEFAULT_CERTIFICATE_POLICY.clone()),
         ..Default::default()
     };
-    let operation = client
-        .create_certificate("certificate-roundtrip", body.try_into()?, None)
-        .await?
-        .into_body()
+    client
+        .begin_create_certificate("certificate-roundtrip", body.try_into()?, None)?
+        .wait()
         .await?;
-    wait_for_certificate_completion(recording, &client, operation).await?;
 
     // Get the latest version of the certificate we just created.
     let certificate = client
@@ -90,15 +88,13 @@ async fn update_certificate_properties(ctx: TestContext) -> Result<()> {
 
     // Create a self-signed certificate.
     let body = CreateCertificateParameters {
-        certificate_policy: Some(DEFAULT_POLICY.clone()),
+        certificate_policy: Some(DEFAULT_CERTIFICATE_POLICY.clone()),
         ..Default::default()
     };
-    let operation = client
-        .create_certificate("update-properties", body.try_into()?, None)
-        .await?
-        .into_body()
+    client
+        .begin_create_certificate("update-properties", body.try_into()?, None)?
+        .wait()
         .await?;
-    wait_for_certificate_completion(recording, &client, operation).await?;
 
     // Get the latest version of the certificate we just created.
     let certificate = client
@@ -153,22 +149,17 @@ async fn list_certificates(ctx: TestContext) -> Result<()> {
     // Create several self-signed certificates.
     let mut names = vec!["list-certificates-1", "list-certificates-2"];
     let body = CreateCertificateParameters {
-        certificate_policy: Some(DEFAULT_POLICY.clone()),
+        certificate_policy: Some(DEFAULT_CERTIFICATE_POLICY.clone()),
         ..Default::default()
     };
-    let operation1 = client
-        .create_certificate("list-certificates-1", body.clone().try_into()?, None)
-        .await?
-        .into_body()
+    client
+        .begin_create_certificate("list-certificates-1", body.clone().try_into()?, None)?
+        .wait()
         .await?;
-    wait_for_certificate_completion(recording, &client, operation1).await?;
-
-    let operation2 = client
-        .create_certificate("list-certificates-2", body.try_into()?, None)
-        .await?
-        .into_body()
+    client
+        .begin_create_certificate("list-certificates-2", body.try_into()?, None)?
+        .wait()
         .await?;
-    wait_for_certificate_completion(recording, &client, operation2).await?;
 
     // List certificates.
     let mut pager = client.list_certificate_properties(None)?.into_stream();
@@ -200,19 +191,17 @@ async fn purge_certificate(ctx: TestContext) -> Result<()> {
 
     // Create a self-signed certificate.
     let body = CreateCertificateParameters {
-        certificate_policy: Some(DEFAULT_POLICY.clone()),
+        certificate_policy: Some(DEFAULT_CERTIFICATE_POLICY.clone()),
         ..Default::default()
     };
-    let operation = client
-        .create_certificate("purge-certificate", body.try_into()?, None)
-        .await?
-        .into_body()
+    const NAME: &str = "purge-certificate";
+    client
+        .begin_create_certificate(NAME, body.try_into()?, None)?
+        .wait()
         .await?;
-    let name = operation.resource_id()?.name;
-    wait_for_certificate_completion(recording, &client, operation).await?;
 
     // Delete the certificate.
-    client.delete_certificate(name.as_ref(), None).await?;
+    client.delete_certificate(NAME.as_ref(), None).await?;
 
     // Because deletes may not happen right away, try purging in a loop.
     let mut retry = match recording.test_mode() {
@@ -221,9 +210,9 @@ async fn purge_certificate(ctx: TestContext) -> Result<()> {
     };
 
     loop {
-        match client.purge_deleted_certificate(name.as_ref(), None).await {
+        match client.purge_deleted_certificate(NAME.as_ref(), None).await {
             Ok(_) => {
-                println!("{name} has been purged");
+                println!("{NAME} has been purged");
                 break;
             }
             Err(err) if matches!(err.http_status(), Some(StatusCode::Conflict)) => {
@@ -279,21 +268,21 @@ async fn sign_jwt_with_ec_certificate(ctx: TestContext) -> Result<()> {
         certificate_policy: Some(policy),
         ..Default::default()
     };
-    let operation = client
-        .create_certificate("ec-certificate-signer", body.try_into()?, None)
-        .await?
-        .into_body()
+    const NAME: &str = "ec-certificate-signer";
+    client
+        .begin_create_certificate(NAME, body.try_into()?, None)?
+        .wait()
         .await?;
-    let ResourceId {
-        vault_url, name, ..
-    } = operation.resource_id()?;
-    wait_for_certificate_completion(recording, &client, operation).await?;
 
     let mut key_options = KeyClientOptions::default();
     recording.instrument(&mut key_options.client_options);
 
     // Sign a JWT.
-    let key_client = KeyClient::new(&vault_url, recording.credential(), Some(key_options))?;
+    let key_client = KeyClient::new(
+        client.endpoint().as_str(),
+        recording.credential(),
+        Some(key_options),
+    )?;
 
     // cspell:disable
     const JWT: &[u8] =
@@ -305,7 +294,7 @@ async fn sign_jwt_with_ec_certificate(ctx: TestContext) -> Result<()> {
         value: Some(digest),
     };
     let signature = key_client
-        .sign(&name, "", body.try_into()?, None)
+        .sign(NAME, "", body.try_into()?, None)
         .await?
         .into_body()
         .await?;
@@ -316,36 +305,84 @@ async fn sign_jwt_with_ec_certificate(ctx: TestContext) -> Result<()> {
     Ok(())
 }
 
-async fn wait_for_certificate_completion(
-    recording: &Recording,
-    client: &CertificateClient,
-    operation: CertificateOperation,
-) -> azure_core::Result<()> {
-    let mut operation = operation;
-    let name = operation.resource_id()?.name;
-    loop {
-        if matches!(operation.status, Some(ref status) if status == "completed") {
-            break;
-        }
+#[recorded::test]
+async fn get_certificate_operation(ctx: TestContext) -> Result<()> {
+    let recording = ctx.recording();
+    recording.remove_sanitizers(&[SANITIZE_BODY_NAME]).await?;
 
-        if let Some(err) = operation.error {
-            return Err(azure_core::Error::new(
-                azure_core::error::ErrorKind::Other,
-                err.message
-                    .unwrap_or_else(|| "failed to create certificate".into()),
-            ));
-        }
+    let mut options = CertificateClientOptions::default();
+    recording.instrument(&mut options.client_options);
 
-        if recording.test_mode() != TestMode::Playback {
-            sleep(Duration::seconds(3)).await;
-        }
+    let client = CertificateClient::new(
+        recording.var("AZURE_KEYVAULT_URL", None).as_str(),
+        recording.credential(),
+        Some(options),
+    )?;
 
-        operation = client
-            .get_certificate_operation(&name, None)
-            .await?
-            .into_body()
-            .await?;
-    }
+    const CERTIFICATE_NAME: &str = "get-certificate-operation";
+
+    // Start creating a self-signed certificate but do not wait until completed.
+    let body = CreateCertificateParameters {
+        certificate_policy: Some(DEFAULT_CERTIFICATE_POLICY.clone()),
+        ..Default::default()
+    };
+    client
+        .begin_create_certificate(CERTIFICATE_NAME, body.try_into()?, None)?
+        // Request not sent until first execution of pipeline.
+        .try_next()
+        .await?;
+
+    // Now get and wait on the pending operation.
+    let operation = client
+        .resume_certificate_operation(CERTIFICATE_NAME, None)?
+        .wait()
+        .await?
+        .into_body()
+        .await?;
+    assert_eq!(operation.status, Some("completed".into()));
+
+    // Get the latest version of the certificate we just created.
+    let certificate = client
+        .get_certificate(CERTIFICATE_NAME, "", None)
+        .await?
+        .into_body()
+        .await?;
+    let version = certificate.resource_id()?.version;
+
+    assert!(certificate.id.is_some());
+    assert!(version.is_some());
+
+    Ok(())
+}
+
+#[recorded::test]
+async fn create_invalid_certificate(ctx: TestContext) -> Result<()> {
+    let recording = ctx.recording();
+    recording.remove_sanitizers(&[SANITIZE_BODY_NAME]).await?;
+
+    let mut options = CertificateClientOptions::default();
+    recording.instrument(&mut options.client_options);
+
+    let client = CertificateClient::new(
+        recording.var("AZURE_KEYVAULT_URL", None).as_str(),
+        recording.credential(),
+        Some(options),
+    )?;
+
+    let body = CreateCertificateParameters {
+        certificate_policy: Some(DEFAULT_CERTIFICATE_POLICY.clone()),
+        ..Default::default()
+    };
+    let err = client
+        .begin_create_certificate("create_invalid_certificate", body.try_into()?, None)?
+        .wait()
+        .await
+        .expect_err("expected HTTP error");
+
+    assert!(matches!(
+        err.kind(),
+        ErrorKind::HttpResponse { status, .. } if *status == StatusCode::BadRequest
+    ));
 
     Ok(())
 }